fraud_detection_module/fraud_detection_system/kafka_stream/requirements.txt
xgboost
numpy
matplotlib
kafka-python
pandas
scikit-learn
joblib
shap